requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-crontab==0.7.1
djangorestframework==3.14.0
gevent==22.10.2
greenlet==2.0.2
idna==3.4
numpy==1.24.2
pandas==1.5.3
pyotp==2.8.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
smartapi-python==1.3.0
sqlparse==0.4.3
urllib3==1.26.14
websocket==0.2.1
zope.event==4.6
zope.interface==5.5.2
